services/api/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.13.3

# AWS SDK -- for SQS and Secrets Manager
boto3==1.35.36

# Metrics -- exposes /metrics endpoint for Prometheus to scrape
prometheus-client==0.21.0

# OpenTelemetry -- distributed tracing
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-sqlalchemy==0.48b0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
